Через Swagger прекрасно работает
