+++ src/main/resources/templates/tour/setting-score.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" xmlns:th="http://www.w3.org/1999/html" xmlns:sec="http://www.w3.org/1999/xhtml">
 <head>
     <meta charset="UTF-8">
     <title>Setting score</title>
@@ -10,6 +10,9 @@
             crossorigin="anonymous"></script>
 </head>
 <body>
+<var allTypes=[1,2,3]></var>
+
+<div th:replace="~{common/navbar :: navi(TourInfo)}"></div>
 <thead>
 <table class="table table-striped table-bordered table table-sm">
     <thead class="table-secondary">
@@ -28,19 +31,55 @@
 
     <tr th:each="raw, iter : ${tournament.getGamesList}">
         <th scope="col" th:text="${iter.count}"></th>
-                <th scope="col" th:text="${tournament.getGamesList.get(iter.count - 1).getPlayerPair.getFirst.firstname + ' ' + tournament.getGamesList.get(iter.count - 1).getPlayerPair.getFirst.lastname}"></th>
-                <th scope="col" th:text="${tournament.getGamesList.get(iter.count - 1).getPlayerPair.getSecond.firstname + ' ' + tournament.getGamesList.get(iter.count - 1).getPlayerPair.getSecond.lastname}"></th>
-        <!--<th th:each="colName : ${tournament.getResultTable.get(iter.count - 1)}" scope="colgroup" width="40px" height="40px" th:text="${colName}"></th>
-        &lt;!&ndash;        <th scope="col" th:text="${raw}"></th>&ndash;&gt;
-        <th scope="col" th:text="${tournament.players.get(iter.count - 1).rating}"></th>
-        <th scope="col" th:text="${tournament.players.get(iter.count - 1).delta}"></th>
-        <th scope="col" th:text="${tournament.players.get(iter.count - 1).sets}"></th>
-        <th scope="col" th:text="${tournament.players.get(iter.count - 1).place}"></th>-->
+        <th scope="col"
+            th:text="${tournament.getGamesList.get(iter.count - 1).getPlayerPair.getFirst.firstname + ' ' + tournament.getGamesList.get(iter.count - 1).getPlayerPair.getFirst.lastname}"></th>
+        <th scope="col"
+            th:text="${tournament.getGamesList.get(iter.count - 1).getPlayerPair.getSecond.firstname + ' ' + tournament.getGamesList.get(iter.count - 1).getPlayerPair.getSecond.lastname}"></th>
+        <th scope="col"
+            th:action="@{/courier/notInTime}" method="post">
+            <select class="form-control" id="playSet1" name="task">
+                <option value="1">1</option>
+                <option value="2">2</option>
+                <option value="3">3</option>
+            </select>
+            <!--                <button type="submit">Not in time</button>-->
+
+        </th>
+
+        <th scope="col"
+        >
+            <select class="form-control" id="playSet2" name="task">
+                <option value="1">1</option>
+                <option value="2">2</option>
+                <option value="3">3</option>
+            </select>
+
+            <!--            <button type="submit" action="'/tour/setscore/0/1/1'">Submit</button>-->
+
+        </th>
+
+        <th scope="col"
+            th:action="@{/courier/notInTime}" method="post">
+            <select class="form-control" id="playSet3" name="task">
+                <option value="1">1</option>
+                <option value="2">2</option>
+                <option value="3">3</option>
+            </select>
+            <!--                <button type="submit">Not in time</button>-->
+
+        </th>
+
         <br>
     </tr>
-    </tbody>
 
 </table>
+
+<td>
+    <a
+       type="button" id="sss" th:href="@{'/tour/setscore/1/1/1'}"
+       class="btn btn-warning">Save</a>
+</td>
+
 </body>
 
 </html>
