+++ src/main/resources/templates/tour/upcoming-tours.html
@@ -73,9 +73,6 @@
                 <tfoot>
                 </tfoot>
             </table>
-        </div>
-
-    </div>
 </div>
 
 <div th:replace="~{common/footer :: footerBlock}"></div>
